requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiomysql==0.2.0
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via redis
bcrypt==4.0.1
    # via -r requirements.in
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==41.0.7
    # via python-jose
dnspython==2.4.2
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==2.1.0.post1
    # via pydantic
exceptiongroup==1.2.0
    # via anyio
fastapi==0.104.1
    # via
    #   -r requirements.in
    #   fastapi-cache2
fastapi-cache2[redis]==0.2.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
passlib==1.7.4
    # via -r requirements.in
pendulum==2.1.2
    # via fastapi-cache2
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic[email]==2.5.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.14.5
    # via pydantic
pymysql==1.1.0
    # via
    #   -r requirements.in
    #   aiomysql
python-dateutil==2.8.2
    # via pendulum
python-dotenv==1.0.0
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
pytzdata==2020.1
    # via pendulum
pyyaml==6.0.1
    # via uvicorn
redis==4.6.0
    # via fastapi-cache2
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uvicorn[standard]==0.24.0.post1
    # via
    #   -r requirements.in
    #   fastapi-cache2
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
